Pronto

+++ app/controllers/links.py
@@ -3,6 +3,7 @@ import string
 
 from app.controllers.validacoes import Validacoes
 from app.models.bd import Connection
+from app.controllers.utils import *
 
 
 class Links:
@@ -11,14 +12,14 @@ class Links:
         self.dominio = "http://127.0.0.1:5000/urls/"
 
 
-    def cadastra_urls(self, URL):
+    def cadastra_urls(self, URL, userid):
         self.shorturl = Links.gerar_codigo(self)
         self.garanta = Validacoes()
         URL = self.garanta.http_existe_na_url(URL)
 
         try:
             self.bd = Connection()
-            self.query = ("INSERT INTO links (hits, url, shorturl) VALUES ({0}, '{1}', '{2}')".format(0, URL, self.shorturl))
+            self.query = ("INSERT INTO links (hits, url, shorturl, usuario_id) VALUES ({0}, '{1}', '{2}', {3})".format(0, URL, self.shorturl, userid))
             self.bd.executarSQL(self.query)
             return self.bd.findOne("SELECT * FROM links WHERE shorturl = '{0}'".format(self.shorturl))
         except:
@@ -96,10 +97,13 @@ class Links:
             self.bd.cursor.close()
             self.bd.connection.close()
 
-    def delete_url(self, id):
+    def delete_url(self, id=None, id_usuario=None):
         try:
+            if id:
+                self.query = ("DELETE FROM links WHERE id = {0}".format(id))
+            if id_usuario:
+                self.query = ("DELETE FROM links WHERE usuario_id = {0}".format(id_usuario))
             self.bd = Connection()
-            self.query = ("DELETE FROM links WHERE id = {0}".format(id))
             self.bd.executarSQL(self.query)
         except:
             print("Erro no deletar registro")
@@ -107,21 +111,39 @@ class Links:
             self.bd.cursor.close()
             self.bd.connection.close()
 
-    def get_analitcs(self):
+    def get_analitcs(self, id=None):
         try:
-            self.bd = Connection()
-            self.query = "select sum(hits), count(url) from links"
-            self.hits_count = self.bd.findOne("select sum(hits), count(url) from links")
-            self.query = "select * from links order by hits desc fetch first 10 rows only"
-            self.top_hits = self.bd.findOne(self.query)
+            if id:
+                self.bd = Connection()
+                self.query = ("select sum(hits), count(url) "
+                              "from links L "
+                              "join users U on L.usuario_id = U.id_usuario "
+                              "where U.id_usuario = '{0}'".format(id))
+                self.hitsUrls = self.bd.findOne(self.query)
 
-            utils = Utils()
-            self.top_hits = utils.gera_dict(self.top_hits)
-
-            self.results = {"urlCount": self.hits_count[0][1],
-                            "hits": self.hits_count[0][0],
-                            "topurls": self.top_hits}
-
+                utils = Utils()
+                self.query = ("select * "
+                              "from links L "
+                              "join users U on L.usuario_id = U.id_usuario "
+                              "where U.id_usuario = '{0}' "
+                              "order by hits desc "
+                              "fetch first 10 rows only".format(id))
+                self.top_hits = self.bd.findOne(self.query)
+                self.top_hits = utils.gera_dict(self.top_hits)
+                self.results = {"urlCount": self.hitsUrls[0][1],
+                                "hits": self.hitsUrls[0][0],
+                                "topurls": self.top_hits}
+            else:
+                self.bd = Connection()
+                self.query = "select sum(hits), count(url) from links"
+                self.hits_count = self.bd.findOne("select sum(hits), count(url) from links")
+                self.query = "select * from links order by hits desc fetch first 10 rows only"
+                self.top_hits = self.bd.findOne(self.query)
+                utils = Utils()
+                self.top_hits = utils.gera_dict(self.top_hits)
+                self.results = {"urlCount": self.hits_count[0][1],
+                                "hits": self.hits_count[0][0],
+                                "topurls": self.top_hits}
 
             return self.results
         except:
@@ -129,3 +151,4 @@ class Links:
         finally:
             self.bd.cursor.close()
             self.bd.connection.close()
+
